fw4: do not allow setting ports for non-TCP or -UDP rules and redirects
Currently, it is possible to set source and destination ports for
'proto' like 'all' or 'icmp', although only 'tcp' and 'udp' ports
are converted into the nftables ruleset.

While this might not be a problem for a protocol like ICMP,
there is a security risk when the user specifies 'all', because
then the rule/redirect applies to all L4 ports, including TCP
and UDP, regardless if the user sets specific ports.

This commit checks for rules and redirects that if any port was
set the protocol must be TCP and/or UDP. If not, the user will
be informed via a warning, and the rule/redirect will be ignored.
Should the protocol be 'all', it is automatically corrected
to TCP and UDP by the ucode function ensure_tcpudp.

Signed-off-by: Til Kaiser <mail@tk154.de>